Moved web socket into chat component refactor message and messages
